Move 'api' package out of 'internal' (#43)
