+++ htsget-server/api/testdata/NA12878.chr20.sample.bam
